Core/Classic UI: Allow to extend expire time for acknowledgements
If the problem was acknowledged and an expire time was set,
check if the newly provided expire time is greater than the old
one. If so, remove the old acknowledgement and create a new one.

The Classic UI shows a new command if the problem has been acknowledged,
additional to the already existing removal command.

Requires the patch in #8445

refs #4908